requirements.txt
asgiref==3.8.1
attrs==23.2.0
Babel==2.14.0
crispy-bootstrap5==2024.2
Django==5.0.3
django-cors-headers==4.4.0
django-countries==7.6
django-crispy-forms==2.1
django-filter==24.2
django-money==3.4.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-simple-history==3.6.0
djangorestframework==3.15.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
phonenumbers==8.13.33
psycopg2-binary==2.9.9
py-moneyed==3.0
referencing==0.35.1
rpds-py==0.18.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
